# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/site-settings/{key}": {
      "get": {
        "operationId": "SiteSettingController_getByKey",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "example": "logo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/GetByKeyResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Setting not found, or private and the caller is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "summary": "Get a site setting by key (e.g. logo, description, slogan)",
        "tags": [
          "SiteSetting"
        ]
      },
      "put": {
        "operationId": "SiteSettingController_upsertByKey",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "example": "logo",
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertByKeyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upserted setting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UpsertByKeyResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create or update a site setting by key",
        "tags": [
          "SiteSetting"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lio API",
    "description": "Authentication and session management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "GetByKeyResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "header"
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "menu": []
            }
          },
          "isPrivate": {
            "type": "boolean",
            "description": "Whether the setting is only returned to authenticated callers",
            "example": false
          }
        },
        "required": [
          "key",
          "data",
          "isPrivate"
        ]
      },
      "UpsertByKeyRequestDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "menu": []
            }
          },
          "isPrivate": {
            "type": "boolean",
            "description": "When true, the setting is only returned to authenticated callers",
            "example": false
          }
        },
        "required": [
          "data"
        ]
      },
      "UpsertByKeyResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "footer"
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "menu": []
            }
          },
          "isPrivate": {
            "type": "boolean",
            "description": "Whether the setting is only returned to authenticated callers",
            "example": false
          }
        },
        "required": [
          "key",
          "data",
          "isPrivate"
        ]
      },
...
    }
  }
}